+++ tests/core/recipes/qchem_recipes/mocked/test_qchem_recipes.py
@@ -3,11 +3,9 @@ from shutil import copy
 
 import pytest
 from ase import units
-from ase.calculators.calculator import FileIOCalculator
 from ase.calculators.lj import LennardJones
 from ase.io import read
 from ase.optimize import FIRE
-from pymatgen.io.ase import AseAtomsAdaptor
 from pymatgen.io.qchem.inputs import QCInput
 
 from quacc import SETTINGS
@@ -96,7 +94,7 @@ def mock_execute3(_self, **kwargs):
 def mock_execute4(self, **kwargs):
     qcin = QCInput.from_file("mol.qin")
     mol = qcin.molecule
-    atoms = AseAtomsAdaptor.get_atoms(mol)
+    atoms = mol.to_ase_atoms()
     atoms.calc = LennardJones()
     atoms.get_potential_energy()
     self.results = atoms.calc.results
@@ -116,7 +114,7 @@ def mock_read(self, **kwargs):
 
 def test_static_job_v1(monkeypatch, tmp_path, test_atoms):
     monkeypatch.chdir(tmp_path)
-    monkeypatch.setattr(FileIOCalculator, "execute", mock_execute1)
+    monkeypatch.setattr(QChem, "execute", mock_execute1)
     charge, spin_multiplicity = check_charge_and_spin(test_atoms)
     output = static_job(test_atoms, charge=charge, spin_multiplicity=spin_multiplicity)
     assert output["atoms"] == test_atoms
@@ -139,7 +137,7 @@ def test_static_job_v1(monkeypatch, tmp_path, test_atoms):
 def test_static_job_v2(monkeypatch, tmp_path, test_atoms):
     monkeypatch.chdir(tmp_path)
 
-    monkeypatch.setattr(FileIOCalculator, "execute", mock_execute2)
+    monkeypatch.setattr(QChem, "execute", mock_execute2)
     charge, spin_multiplicity = check_charge_and_spin(test_atoms, charge=-1)
     output = static_job(
         test_atoms,
@@ -169,7 +167,7 @@ def test_static_job_v2(monkeypatch, tmp_path, test_atoms):
 def test_static_job_v3(monkeypatch, tmp_path, test_atoms):
     monkeypatch.chdir(tmp_path)
 
-    monkeypatch.setattr(FileIOCalculator, "execute", mock_execute3)
+    monkeypatch.setattr(QChem, "execute", mock_execute3)
     charge, spin_multiplicity = check_charge_and_spin(test_atoms)
     output = static_job(
         test_atoms,
@@ -196,7 +194,7 @@ def test_static_job_v3(monkeypatch, tmp_path, test_atoms):
 def test_static_job_v4(monkeypatch, tmp_path, os_atoms):
     monkeypatch.chdir(tmp_path)
     monkeypatch.setattr(QChem, "read_results", mock_read)
-    monkeypatch.setattr(FileIOCalculator, "execute", mock_execute4)
+    monkeypatch.setattr(QChem, "execute", mock_execute4)
     charge, spin_multiplicity = check_charge_and_spin(os_atoms)
     assert static_job(os_atoms, charge=charge, spin_multiplicity=spin_multiplicity)
 
@@ -217,7 +215,7 @@ def test_static_job_v5(tmp_path, monkeypatch, test_atoms):
 def test_relax_job_v1(monkeypatch, tmp_path, test_atoms):
     monkeypatch.chdir(tmp_path)
 
-    monkeypatch.setattr(FileIOCalculator, "execute", mock_execute1)
+    monkeypatch.setattr(QChem, "execute", mock_execute1)
     charge, spin_multiplicity = check_charge_and_spin(test_atoms)
     output = relax_job(
         test_atoms,
@@ -246,7 +244,7 @@ def test_relax_job_v1(monkeypatch, tmp_path, test_atoms):
 @pytest.mark.skipif(sella is None, reason="Does not have Sella")
 def test_relax_job_v2(monkeypatch, tmp_path, test_atoms):
     monkeypatch.chdir(tmp_path)
-    monkeypatch.setattr(FileIOCalculator, "execute", mock_execute2)
+    monkeypatch.setattr(QChem, "execute", mock_execute2)
     charge, spin_multiplicity = check_charge_and_spin(test_atoms, charge=-1)
     output = relax_job(
         test_atoms,
@@ -277,7 +275,7 @@ def test_relax_job_v2(monkeypatch, tmp_path, test_atoms):
 @pytest.mark.skipif(sella is None, reason="Does not have Sella")
 def test_relax_job_v3(monkeypatch, tmp_path, test_atoms):
     monkeypatch.chdir(tmp_path)
-    monkeypatch.setattr(FileIOCalculator, "execute", mock_execute3)
+    monkeypatch.setattr(QChem, "execute", mock_execute3)
     charge, spin_multiplicity = check_charge_and_spin(test_atoms)
     output = relax_job(
         test_atoms,
@@ -313,7 +311,7 @@ def test_relax_job_v4(tmp_path, monkeypatch, test_atoms):
 
 def test_freq_job_v1(monkeypatch, tmp_path, test_atoms):
     monkeypatch.chdir(tmp_path)
-    monkeypatch.setattr(FileIOCalculator, "execute", mock_execute5)
+    monkeypatch.setattr(QChem, "execute", mock_execute5)
     charge, spin_multiplicity = check_charge_and_spin(test_atoms, charge=-1)
     output = freq_job(
         test_atoms,
@@ -349,7 +347,7 @@ def test_freq_job_v1(monkeypatch, tmp_path, test_atoms):
 def test_ts_job_v1(monkeypatch, tmp_path, test_atoms):
     monkeypatch.chdir(tmp_path)
 
-    monkeypatch.setattr(FileIOCalculator, "execute", mock_execute1)
+    monkeypatch.setattr(QChem, "execute", mock_execute1)
     charge, spin_multiplicity = check_charge_and_spin(test_atoms)
     output = ts_job(
         test_atoms,
@@ -377,7 +375,7 @@ def test_ts_job_v1(monkeypatch, tmp_path, test_atoms):
 @pytest.mark.skipif(sella is None, reason="Does not have Sella")
 def test_ts_job_v2(monkeypatch, tmp_path, test_atoms):
     monkeypatch.chdir(tmp_path)
-    monkeypatch.setattr(FileIOCalculator, "execute", mock_execute2)
+    monkeypatch.setattr(QChem, "execute", mock_execute2)
     charge, spin_multiplicity = check_charge_and_spin(test_atoms, charge=-1)
     output = ts_job(
         test_atoms,
@@ -408,7 +406,7 @@ def test_ts_job_v2(monkeypatch, tmp_path, test_atoms):
 @pytest.mark.skipif(sella is None, reason="Does not have Sella")
 def test_ts_job_v3(monkeypatch, tmp_path, test_atoms):
     monkeypatch.chdir(tmp_path)
-    monkeypatch.setattr(FileIOCalculator, "execute", mock_execute3)
+    monkeypatch.setattr(QChem, "execute", mock_execute3)
     charge, spin_multiplicity = check_charge_and_spin(test_atoms)
     output = ts_job(
         test_atoms,
@@ -456,7 +454,7 @@ def test_irc_job_v1(monkeypatch, tmp_path, test_atoms):
     monkeypatch.chdir(tmp_path)
 
     monkeypatch.setattr(QChem, "read_results", mock_read)
-    monkeypatch.setattr(FileIOCalculator, "execute", mock_execute4)
+    monkeypatch.setattr(QChem, "execute", mock_execute4)
 
     charge, spin_multiplicity = check_charge_and_spin(test_atoms)
     output = irc_job(
@@ -548,7 +546,7 @@ def test_quasi_irc_job(monkeypatch, tmp_path, test_atoms):
     monkeypatch.chdir(tmp_path)
 
     monkeypatch.setattr(QChem, "read_results", mock_read)
-    monkeypatch.setattr(FileIOCalculator, "execute", mock_execute4)
+    monkeypatch.setattr(QChem, "execute", mock_execute4)
 
     charge, spin_multiplicity = check_charge_and_spin(test_atoms)
     output = quasi_irc_job(