+++ tests/core/recipes/vasp_recipes/mocked/test_vasp_recipes.py
@@ -2,7 +2,7 @@ import pytest
 from ase.build import bulk, molecule
 
 from quacc import SETTINGS
-from quacc.recipes.vasp.core import double_relax_job, relax_job, static_job
+from quacc.recipes.vasp.core import double_relax_flow, relax_job, static_job
 from quacc.recipes.vasp.mp import (
     mp_gga_relax_flow,
     mp_gga_relax_job,
@@ -98,12 +98,12 @@ def test_relax_job(tmp_path, monkeypatch):
     assert output["parameters"]["isif"] == 2
 
 
-def test_doublerelax_job(tmp_path, monkeypatch):
+def test_doublerelax_flow(tmp_path, monkeypatch):
     monkeypatch.chdir(tmp_path)
 
     atoms = bulk("Al")
 
-    output = double_relax_job(atoms)
+    output = double_relax_flow(atoms)
     assert output["relax1"]["nsites"] == len(atoms)
     assert output["relax1"]["parameters"]["isym"] == 0
     assert output["relax1"]["parameters"]["nsw"] > 0
@@ -117,7 +117,7 @@ def test_doublerelax_job(tmp_path, monkeypatch):
     assert output["relax2"]["parameters"]["lwave"] is False
     assert output["relax2"]["parameters"]["encut"] == 520
 
-    output = double_relax_job(atoms, relax2_kwargs={"nelmin": 6})
+    output = double_relax_flow(atoms, relax2_kwargs={"nelmin": 6})
     assert output["relax1"]["nsites"] == len(atoms)
     assert output["relax1"]["parameters"]["isym"] == 0
     assert output["relax1"]["parameters"]["nsw"] > 0
@@ -132,7 +132,7 @@ def test_doublerelax_job(tmp_path, monkeypatch):
     assert output["relax2"]["parameters"]["encut"] == 520
     assert output["relax2"]["parameters"]["nelmin"] == 6
 
-    output = double_relax_job(atoms, relax_cell=False)
+    output = double_relax_flow(atoms, relax_cell=False)
     assert output["relax1"]["nsites"] == len(atoms)
     assert output["relax1"]["parameters"]["isym"] == 0
     assert output["relax1"]["parameters"]["nsw"] > 0
@@ -146,7 +146,7 @@ def test_doublerelax_job(tmp_path, monkeypatch):
     assert output["relax2"]["parameters"]["encut"] == 520
     assert output["relax2"]["parameters"]["isif"] == 2
 
-    assert double_relax_job(atoms, relax1_kwargs={"kpts": [1, 1, 1]})
+    assert double_relax_flow(atoms, relax1_kwargs={"kpts": [1, 1, 1]})
 
 
 def test_slab_static_job(tmp_path, monkeypatch):
@@ -424,9 +424,7 @@ def test_mp_metagga_relax_job(tmp_path, monkeypatch):
 
     atoms = bulk("Al")
 
-    output = mp_metagga_relax_job(atoms)
-    assert output["relax2"]["nsites"] == len(atoms)
-    assert output["relax2"]["parameters"] == {
+    ref_parameters = {
         "algo": "all",
         "ediff": 1e-5,
         "ediffg": -0.02,
@@ -456,26 +454,32 @@ def test_mp_metagga_relax_job(tmp_path, monkeypatch):
         "pp": "pbe",
         "setups": {"Al": ""},
     }
+    ref_parameters2 = ref_parameters.copy()
+    ref_parameters2["magmom"] = [0.0]
+
+    output = mp_metagga_relax_job(atoms)
+    assert output["parameters"] == ref_parameters
+    assert output["nsites"] == len(atoms)
 
     output = mp_metagga_relax_job(atoms, bandgap=0)
-    assert output["relax2"]["nsites"] == len(atoms)
-    assert output["relax2"]["parameters"]["metagga"].lower() == "r2scan"
-    assert output["relax2"]["parameters"]["ediffg"] == -0.02
-    assert output["relax2"]["parameters"]["encut"] == 680
-    assert output["relax2"]["parameters"]["kspacing"] == 0.22
-    assert output["relax2"]["parameters"]["ismear"] == 0
-    assert output["relax2"]["parameters"]["sigma"] == 0.05
-    assert output["relax2"]["parameters"]["pp"] == "pbe"
+    assert output["nsites"] == len(atoms)
+    assert output["parameters"]["metagga"].lower() == "r2scan"
+    assert output["parameters"]["ediffg"] == -0.02
+    assert output["parameters"]["encut"] == 680
+    assert output["parameters"]["kspacing"] == 0.22
+    assert output["parameters"]["ismear"] == 0
+    assert output["parameters"]["sigma"] == 0.05
+    assert output["parameters"]["pp"] == "pbe"
 
     output = mp_metagga_relax_job(atoms, bandgap=100)
-    assert output["relax2"]["nsites"] == len(atoms)
-    assert output["relax2"]["parameters"]["metagga"].lower() == "r2scan"
-    assert output["relax2"]["parameters"]["ediffg"] == -0.02
-    assert output["relax2"]["parameters"]["encut"] == 680
-    assert output["relax2"]["parameters"]["kspacing"] == 0.44
-    assert output["relax2"]["parameters"]["ismear"] == 0
-    assert output["relax2"]["parameters"]["sigma"] == 0.05
-    assert output["relax2"]["parameters"]["pp"] == "pbe"
+    assert output["nsites"] == len(atoms)
+    assert output["parameters"]["metagga"].lower() == "r2scan"
+    assert output["parameters"]["ediffg"] == -0.02
+    assert output["parameters"]["encut"] == 680
+    assert output["parameters"]["kspacing"] == 0.44
+    assert output["parameters"]["ismear"] == 0
+    assert output["parameters"]["sigma"] == 0.05
+    assert output["parameters"]["pp"] == "pbe"
 
 
 def test_mp_metagga_static_job(tmp_path, monkeypatch):
@@ -521,13 +525,13 @@ def test_mp_metagga_relax_flow(tmp_path, monkeypatch):
 
     output = mp_metagga_relax_flow(atoms)
     assert output["static"]["nsites"] == len(atoms)
-    assert output["relax"]["relax2"]["parameters"]["metagga"].lower() == "r2scan"
-    assert output["relax"]["relax2"]["parameters"]["ediffg"] == -0.02
-    assert output["relax"]["relax2"]["parameters"]["encut"] == 680
-    assert output["relax"]["relax2"]["parameters"]["ismear"] == 0
-    assert output["relax"]["relax2"]["parameters"]["sigma"] == 0.05
-    assert output["relax"]["relax2"]["parameters"]["kspacing"] == 0.22
-    assert output["relax"]["relax2"]["parameters"]["pp"] == "pbe"
+    assert output["relax2"]["parameters"]["metagga"].lower() == "r2scan"
+    assert output["relax2"]["parameters"]["ediffg"] == -0.02
+    assert output["relax2"]["parameters"]["encut"] == 680
+    assert output["relax2"]["parameters"]["ismear"] == 0
+    assert output["relax2"]["parameters"]["sigma"] == 0.05
+    assert output["relax2"]["parameters"]["kspacing"] == 0.22
+    assert output["relax2"]["parameters"]["pp"] == "pbe"
     assert output["prerelax"]["parameters"]["gga"] == "ps"
     assert output["prerelax"]["parameters"]["ismear"] == 0
     assert output["prerelax"]["parameters"]["pp"] == "pbe"
@@ -535,14 +539,14 @@ def test_mp_metagga_relax_flow(tmp_path, monkeypatch):
     atoms = bulk("C")
     output = mp_metagga_relax_flow(atoms)
     assert output["static"]["nsites"] == len(atoms)
-    assert output["relax"]["relax2"]["parameters"]["metagga"].lower() == "r2scan"
-    assert output["relax"]["relax2"]["parameters"]["ediffg"] == -0.02
-    assert output["relax"]["relax2"]["parameters"]["encut"] == 680
-    assert output["relax"]["relax2"]["parameters"]["ismear"] == 0
-    assert output["relax"]["relax2"]["parameters"]["kspacing"] == pytest.approx(
+    assert output["relax2"]["parameters"]["metagga"].lower() == "r2scan"
+    assert output["relax2"]["parameters"]["ediffg"] == -0.02
+    assert output["relax2"]["parameters"]["encut"] == 680
+    assert output["relax2"]["parameters"]["ismear"] == 0
+    assert output["relax2"]["parameters"]["kspacing"] == pytest.approx(
         0.28329488761304206
     )
-    assert output["relax"]["relax2"]["parameters"]["pp"] == "pbe"
+    assert output["relax2"]["parameters"]["pp"] == "pbe"
     assert output["prerelax"]["parameters"]["ismear"] == 0
     assert output["prerelax"]["parameters"]["pp"] == "pbe"
 
@@ -554,14 +558,14 @@ def test_mp_metagga_relax_flow(tmp_path, monkeypatch):
     assert output["static"]["parameters"]["ismear"] == -5
     assert output["static"]["parameters"]["nsw"] == 0
     assert output["static"]["parameters"]["algo"] == "fast"
-    assert output["relax"]["relax2"]["parameters"]["metagga"].lower() == "r2scan"
-    assert output["relax"]["relax2"]["parameters"]["ediffg"] == -0.02
-    assert output["relax"]["relax2"]["parameters"]["encut"] == 680
-    assert output["relax"]["relax2"]["parameters"]["ismear"] == 0
-    assert output["relax"]["relax2"]["parameters"]["kspacing"] == pytest.approx(
+    assert output["relax2"]["parameters"]["metagga"].lower() == "r2scan"
+    assert output["relax2"]["parameters"]["ediffg"] == -0.02
+    assert output["relax2"]["parameters"]["encut"] == 680
+    assert output["relax2"]["parameters"]["ismear"] == 0
+    assert output["relax2"]["parameters"]["kspacing"] == pytest.approx(
         0.28329488761304206
     )
-    assert output["relax"]["relax2"]["parameters"]["pp"] == "pbe"
+    assert output["relax2"]["parameters"]["pp"] == "pbe"
     assert output["prerelax"]["parameters"]["ismear"] == 0
     assert output["prerelax"]["parameters"]["pp"] == "pbe"
 
@@ -570,10 +574,9 @@ def test_mp_gga_relax_job():
     atoms = bulk("Ni") * (2, 1, 1)
     atoms[0].symbol = "O"
     output = mp_gga_relax_job(atoms)
-    assert output["relax2"]["nsites"] == len(atoms)
-    assert output["relax1"]["atoms"].get_chemical_symbols() == ["O", "Ni"]
-    assert output["relax2"]["atoms"].get_chemical_symbols() == ["O", "Ni"]
-    assert output["relax2"]["parameters"] == {
+
+    assert output["nsites"] == len(atoms)
+    assert output["parameters"] == {
         "algo": "fast",
         "ediff": 0.0001,
         "efermi": "midgap",  # added by copilot
@@ -603,6 +606,7 @@ def test_mp_gga_relax_job():
         "pp": "pbe",
         "setups": {"O": "", "Ni": "_pv"},
     }
+    assert output["atoms"].get_chemical_symbols() == ["O", "Ni"]
 
 
 def test_mp_gga_static_job():
@@ -676,8 +680,8 @@ def test_mp_gga_relax_flow():
         "pp": "pbe",
         "setups": {"O": "", "Ni": "_pv"},
     }
-    assert output["relax"]["relax1"]["parameters"] == relax_params
-    assert output["relax"]["relax2"]["parameters"] == relax_params
+    assert output["relax1"]["parameters"] == relax_params
+    assert output["relax2"]["parameters"] == relax_params
     assert output["static"]["parameters"] == {
         "algo": "fast",
         "ediff": 0.0001,
@@ -707,3 +711,15 @@ def test_mp_gga_relax_flow():
         "pp": "pbe",
         "setups": {"Ni": "_pv", "O": ""},
     }
+
+
+def test_mp_relax_flow_custom():
+    atoms = bulk("Ni") * (2, 1, 1)
+    atoms[0].symbol = "O"
+    output = mp_metagga_relax_flow(
+        mp_gga_relax_flow(atoms, job_params={"mp_gga_relax_job": {"nsw": 0}})["static"][
+            "atoms"
+        ],
+        job_params={"mp_metagga_relax_job": {"nsw": 0}},
+    )
+    assert output["relax2"]["parameters"]["nsw"] == 0